requirements.txt
py-cord
youtube-dl
PyNaCl
asyncpraw
lyrics-extractor
wikipedia
google-api-python-client
psycopg2-binary
pillow
qrcode
imgurpython
python-dotenv
wheel
scrapetube
googletrans==4.0.0rc1